Added an ontrigger enter
